Rename temp_objectpack_texture_replace_hack.py
